Add scripts to handle dockerising project
The dockerfile contains instructions to create a
openhim-mediator-file-queue mediator docker image. The ignore file will
prevent unnecessary components being added to the context thereby
reducing the image size.
Dockerising this project is part of a larger effort to improve the
infrastructure of the MomConnect project though this component is
generic.

MOM-544