Remove environment variable ENDPOINTS
This field can be configured from the OpenHIM Console. PReviously it was
configured via a JSON file.

This was changed in favour of the ease of environment variables. The
specific ENDPOINT field support will be dropped as the field is
redundant in initial configuration as the ENDPOINT value is also stored
in the mediator.json file. The value is sent from the mediator to the
OpenHIM. The OpenHIM then sends this field as is or updated via the
OpenHIM Console back to the file-queue-mediator which then updates teh
apiCOnfig option - htereby overwriting whatever value previously existed
there that would have originated from the Environment variable ENDPOINTS

MOM-554