Remove weird logging function reassign
I don't know what this code was intended to do but it does not appear to
have an effect when removed.

Reassigning the console.log and console.error methods to a blank
function seems very wrong... If you don't log and error messages coming
through to the console rather make use of the log level.

MOM-554