Remove dangerous exit code
This is a piece of very poorly thought out code.
It would have caused the app to exit if an error occured whilst
updating (in the greater scheme of things) unimportant config...
If ANY kind of error occurred!!!
What the stupid shit...

Just logging the error message now - which isn't amazing but until we
have time to rewrite thiss mediator it will have to do.

MOM-554