Replace the FS.rename method with npm module 'mv'
The mv module by default uses the fs.rename method but falls back to an alternative method of copy/unlink to ensure the same functionality occurs
The reason for this change is when using this service within docker, you want to mount the queue directories within the host system instead of within the container. If this is mounted within the container and the container is destroyed, you will lose all the documents in the queue.
https://stackoverflow.com/questions/43206198/what-does-the-exdev-cross-device-link-not-permitted-error-mean

Changing the docker storage engine could also solve this if setup correctly
https://dker.ru/docs/docker-engine/user-guide/docker-storage-drivers/overlayfs-storage-in-practice/#container-reads-and-writes-with-overlay

MOM-544